+++ TARge21House/TARge21House/Models/House/HouseIndexViewModel.cs
@@ -5,7 +5,7 @@
 		public Guid? Id { get; set; }
 		public int Bedrooms { get; set; }
 		public int Bathrooms { get; set; }
-		public double RentPrice { get; set; }
+		public int RentPrice { get; set; }
 		public DateTime BuiltDate { get; set; }
 	}
 }
